requirements.txt
streamlit
tensorflow
opencv-python
numpy
pillow
requests
beautifulsoup4
pandas
lxml
regex
psycopg2-binary
streamlit-cookies-manager
transformers
torch
torchvision
kaggle
datasets
tf-keras